Refactor crafting overlay into module
